+++ src/screens/general/SubscriberOwnerDashboard.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { PayPalButton } from 'react-paypal-button-v2';
 import '../../styles/SubscriberOwnerDashboard.css';
 import { getLoggedInUserId } from '../../lib/auth';
-import { centsToDollars, createPayment } from '../../lib/subscriberHelper';
+import { centsToDollars } from '../../lib/subscriberHelper';
+import recordPaymentSuccess from '../../lib/paypal';
 
 import secret from '../../secret';
 
@@ -15,7 +16,7 @@ export default class SubscriberOwnerDashboard extends React.Component {
     this.state = {
       latestBill: bills.filter(bill => bill['Is Latest'])[0]
     };
-    this.onSuccess = this.onSuccess.bind(this);
+    this.onPaypalPaymentSuccess = this.onPaypalPaymentSuccess.bind(this);
   }
 
   componentDidMount() {
@@ -27,119 +28,9 @@ export default class SubscriberOwnerDashboard extends React.Component {
     }
   }
 
-  onSuccess(details, data) {
-    /*
-        DETAILS FIELDS
-  
-        create_time: "2019-11-12T04:44:42Z"
-        id: "SOME_ID_HERE"
-        intent: "SOME INTENT HERE"
-        links: [{
-          href: "SOME HREF HERE"
-          method: "HTTP METHOD"
-          rel: "self"
-          title: "SOME TITLE"
-        }]
-        payer: {
-          email_address: "SOME EMAIL ADDRESS", 
-          payer_id: "SOME PAYER ID", 
-          address: {…}, 
-          name: {…}
-        }
-        purchase_units: [{…(EXPANDED BELOW)}]
-        status: "SOME STATUS"
-        update_time: "2019-11-12T04:45:33Z"
-  
-        ----
-  
-        purchase_units (EXPANDED)
-  
-        amount: 
-            value: "SOME VALUE (IN DOLLARS)", 
-            currency_code: "SOME CURRENCY CODE"
-        payee: 
-            email_address: "SOME EMAIL ADDREESS", 
-            merchant_id: "SOME MERCHANT ID"
-        payments: 
-            captures: [{
-                amount: {value: "SOME VALUE (IN DOLLARS)", currency_code: "SOME CURRENCY CODE"}
-                create_time: "2019-11-12T04:58:14Z"
-                final_capture: true/false
-                id: "SOME ID"
-                links: (3) [{…}, {…}, {…}]
-                seller_protection: {status: "SOME STATUS", dispute_categories: Array(2)}
-                status: "SOME STATUS"
-                update_time: "2019-11-12T04:58:14Z"
-            }]
-              
-            reference_id: "SOME REFERENCE ID"
-        shipping:
-            address:
-                address_line_1: "X X STREET"
-                admin_area_1: "SOME STATE"
-                admin_area_2: "SOME AREA"
-                country_code: "SOME COUNTRY"
-                postal_code: "POSTAL CODE HERE"
-            name:
-                full_name: "FULL NAME HERE"
-        
-        status: "SOME STATUS HERE"
-        update_time: "2019-11-12T04:58:14Z"
-     */
-
-    /*
-        DATA FIELDS
-        
-        orderID: "SOME ORDER ID"
-        payerID: "SOME PAYER ID"
-     
-     */
-
+  onPaypalPaymentSuccess(details, data) {
     const { latestBill } = this.state;
-    const owner = latestBill['Subscriber Owner'];
-    const bill = latestBill.ID;
-
-    const { orderID, payerID } = data;
-
-    const { id, intent, status, payer } = details;
-    const { amount, shipping } = details.purchase_units[0]; // assumes purchase_units is only of length 1
-
-    if (details.purchase_units.length > 1) {
-      console.warn('length of details.purchase_units > 1');
-    }
-
-    const { address, name } = shipping;
-    const addressToSave = `${address.address_line_1}, ${address.admin_area_2} ${address.country_code} ${address.admin_area_1} ${address.postal_code}`;
-
-    const record = {
-      fields: {
-        Owner: [owner],
-        Status: status,
-        'Subscriber Bill': [bill],
-        'Order ID': orderID,
-        'Payer ID': payerID,
-        Amount: parseFloat(amount.value, 10) * 100,
-        'Currency Code': amount.currency_code,
-        Address: addressToSave,
-        'Payer Full Name': name.full_name,
-        'Payer Email': payer.email_address,
-        Intent: intent,
-        'Payment Create Time': details.create_time,
-        'Payment Update Time': details.update_time
-      }
-    };
-
-    if (id !== orderID) {
-      record.fields.Notes = 'id and order id mismatch';
-    }
-
-    createPayment(record)
-      .then(paymentId => {
-        console.log(paymentId);
-      })
-      .catch(err => {
-        console.error(err);
-      });
+    recordPaymentSuccess(details, data, latestBill);
   }
 
   render() {
@@ -184,7 +75,7 @@ export default class SubscriberOwnerDashboard extends React.Component {
                 <br />
                 <PayPalButton
                   amount={amtDue}
-                  onSuccess={this.onSuccess}
+                  onSuccess={this.onPaypalPaymentSuccess}
                   options={{
                     clientId
                   }}
